feat: add config
